+++ Multi-form/Multi-form/FormView.swift
@@ -1,71 +1,25 @@
 import SwiftUI
 
-
-struct Theme: Identifiable {
-    let type: ThemeType
-    var id: String { type.title }
-    var isCheck: Bool
-}
-
-enum ThemeType {
-    case meal
-    case meeting
-    case travel
-    case etc
-    
-    var title: String {
-        switch self {
-        case .meal:
-            return "식사 약속"
-        case .meeting:
-            return "미팅 약속"
-        case .travel:
-            return "여행 약속"
-        case .etc:
-            return "기타 약속"
-        }
-    }
-}
-
-struct ThemeCellView: View {
-    @Binding var theme: Theme
-    
-    var body: some View {
-        HStack {
-            Text(theme.type.title)
-                .foregroundColor(.gray)
-            Spacer()
-            Image(systemName:
-                    theme.isCheck ? "checkmark.circle.fill" : "checkmark.circle")
-                .foregroundColor(.gray)
-        }
-        .padding()
-        .background(Color(red: 0.94,green: 0.94,blue: 0.94))
-        .cornerRadius(10)
-        .listRowSeparator(.hidden)
-        .onTapGesture {
-            theme.isCheck.toggle()
-        }
-    }
-}
-
-
-
 struct FormView: View {
     @Environment(\.dismiss) var dismiss
     
-    @State private var themes = [
-        Theme(type: .meal, isCheck: false),
-        Theme(type: .meeting, isCheck: false),
-        Theme(type: .travel, isCheck: false),
-        Theme(type: .etc, isCheck: false)
+    @State var currentPageIndex = 0
+    
+    /// isVerify 가 true 인 경우, '다음' 버튼 활성화
+    /// 1 페이지: 하나 체크한 경우
+    /// 2 페이지: 텍스트필드 2개에 값이 모두 있는 경우
+    @State var isVerify = true
+    
+    private let titleList = [
+        "약속 테마를 선택해 주세요!",
+        "약속명과 장소를 입력해 주세요!"
     ]
     
     var body: some View {
         HStack {
             VStack {
                 HStack {
-                    Text("1/5")
+                    Text("\(currentPageIndex + 1)/5")
                     Spacer()
                     Button {
                         dismiss()
@@ -74,19 +28,24 @@ struct FormView: View {
                      }
                 }
                 .padding(.bottom)
-                Text("약속 테마를 선택해 주세요!")
+                Text(titleList[currentPageIndex])
                     .frame(maxWidth: .infinity, alignment: .leading)
                     .padding(.bottom)
-                List($themes) { $theme in
-                    ThemeCellView(theme: $theme)
+                // MARK: - 하위뷰
+                TabView {
+                    ThemeFormView()
+                    ThemeFormView()
                 }
-                .listStyle(.plain)
-                .padding(EdgeInsets(top: -10, leading: -20, bottom: -10, trailing: -20))
-
+                .tabViewStyle(.page(indexDisplayMode: .never))
+                // ----
                 Spacer()
-                Text("Check Count = \(themes.filter{ $0.isCheck }.count)")
+                
                 Button {
                     print("next button")
+                    // TODO: -
+                    // 다음 버튼을 누르면 현재 페이지인덱스 += 1
+                    // 이전 버튼을 누르면 현재 페이지인덱스 -= 1
+                    currentPageIndex = 1
                 } label: {
                     Text("다음")
                         .padding()
